@@ -1,33 +1,47 @@
 {
-  "name": "rampvis-ui",
-  "version": "1.0.0",
+  "private": true,
   "scripts": {
-    "dev": "env-cmd -f .env.development next dev",
-    "dev-prod": "env-cmd -f .env.production next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "build": "next build && next export",
+    "dev": "next dev",
+    "fix:eslint": "next lint --fix",
+    "fix:prettier": "prettier --write \"**/*\"",
+    "fix:yarn-deduplicate": "yarn install; yarn-deduplicate --strategy=fewer; yarn install",
+    "lint:eslint": "next lint",
+    "lint:prettier": "prettier --check \"**/*\"",
+    "lint:tsc": "yarn tsc --noEmit",
+    "lint:yarn-deduplicate": "yarn-deduplicate --fail --list --strategy=fewer",
+    "prepare": "husky install",
+    "start": "next start"
+  },
+  "lint-staged": {
+    "**/*": [
+      "prettier --write"
+    ]
   },
   "dependencies": {
-    "@emotion/cache": "^11.4.0",
-    "@emotion/react": "^11.4.1",
+    "@emotion/cache": "^11.5.0",
+    "@emotion/react": "^11.5.0",
     "@emotion/styled": "^11.3.0",
     "@material-ui/core": "5.0.0-alpha.34",
     "@material-ui/icons": "5.0.0-alpha.34",
     "@material-ui/lab": "5.0.0-alpha.34",
     "@material-ui/system": "5.0.0-alpha.34",
+    "@mui/icons-material": "^5.0.1",
+    "@mui/lab": "5.0.0-alpha.53",
+    "@mui/material": "^5.0.6",
+    "@mui/styles": "^5.0.2",
     "@zeit/next-css": "^1.0.1",
-    "axios": "^0.21.1",
+    "axios": "^0.24.0",
     "d3": "^5.16.0",
-    "date-fns": "^2.23.0",
+    "date-fns": "^2.25.0",
     "formik": "^2.2.9",
     "history": "^5.0.1",
-    "i18next": "^20.4.0",
+    "i18next": "^21.3.3",
     "jquery": "^3.6.0",
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
-    "next": "^11.1.0",
+    "next": "^12.0.1",
     "nprogress": "^0.2.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
@@ -35,57 +49,43 @@
     "react-beautiful-dnd": "^13.1.0",
     "react-dom": "^17.0.2",
     "react-draft-wysiwyg": "^1.14.7",
-    "react-dropzone": "^11.3.4",
-    "react-helmet-async": "^1.0.9",
-    "react-hot-toast": "^2.1.0",
-    "react-i18next": "^11.11.4",
-    "react-markdown": "^7.0.0",
+    "react-dropzone": "^11.4.2",
+    "react-helmet-async": "^1.1.2",
+    "react-hot-toast": "^2.1.1",
+    "react-i18next": "^11.13.0",
+    "react-markdown": "^7.1.0",
     "react-modal-image": "^2.5.0",
     "react-perfect-scrollbar": "^1.5.8",
     "react-quill": "^1.3.5",
-    "react-redux": "^7.2.4",
-    "react-router": "^6.0.0-beta.0",
+    "react-redux": "^7.2.6",
     "react-syntax-highlighter": "^15.4.4",
-    "stylis-plugin-rtl": "^2.1.0",
-    "yup": "^0.32.9"
+    "stylis-plugin-rtl": "^2.1.1",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
     "@types/d3": "^5.16.0",
-    "@types/jquery": "^3.5.6",
-    "@types/lodash": "^4.14.172",
+    "@types/jquery": "^3.5.8",
+    "@types/lodash": "^4.14.176",
     "@types/nprogress": "^0.2.0",
-    "@types/react": "^17.0.19",
-    "@types/react-beautiful-dnd": "^13.1.1",
-    "@types/react-dom": "^17.0.9",
+    "@types/react": "^17.0.33",
+    "@types/react-beautiful-dnd": "^13.1.2",
+    "@types/react-dom": "^17.0.10",
     "@types/react-draft-wysiwyg": "^1.13.3",
     "@types/yup": "^0.29.13",
-    "env-cmd": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-typescript": "^12.3.1",
+    "@typescript-eslint/eslint-plugin": "^5.2.0",
+    "eslint": "<8.0.0",
+    "eslint-config-next": "12.0.1",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.24.0",
+    "eslint-plugin-import": "^2.25.2",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.24.0",
+    "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-config-next": "11.1.0",
-    "prettier": "^2.3.2",
-    "typescript": "^4.2.4"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "husky": "^7.0.4",
+    "lint-staged": "^11.2.6",
+    "prettier": "^2.4.1",
+    "prettier-plugin-packagejson": "^2.2.13",
+    "prettier-plugin-sh": "^0.7.1",
+    "typescript": "^4.2.4",
+    "yarn-deduplicate": "^3.1.0"
   }
 }
