@@ -1,21 +1,22 @@
 import { FC, useState } from "react";
-import { makeStyles, IconButton } from "@material-ui/core";
-import Paper from "@material-ui/core/Paper";
-import Table from "@material-ui/core/Table";
-import TableBody from "@material-ui/core/TableBody";
-import TableCell from "@material-ui/core/TableCell";
-import TableContainer from "@material-ui/core/TableContainer";
-import TableHead from "@material-ui/core/TableHead";
-import TablePagination from "@material-ui/core/TablePagination";
-import TableRow from "@material-ui/core/TableRow";
-import { blue } from "@material-ui/core/colors";
+import { Box } from "@mui/material";
+import { makeStyles } from "@mui/styles";
+import Paper from "@mui/material/Paper";
+import Table from "@mui/material/Table";
+import TableBody from "@mui/material/TableBody";
+import TableCell from "@mui/material/TableCell";
+import TableContainer from "@mui/material/TableContainer";
+import TableHead from "@mui/material/TableHead";
+import TablePagination from "@mui/material/TablePagination";
+import TableRow from "@mui/material/TableRow";
+import { blue } from "@mui/material/colors";
+import CircularProgress from "@mui/material/CircularProgress";
 import { useRouter } from "next/router";
-import LinkIcon from "@material-ui/icons/Link";
-
+import Router from "next/router";
 import useSettings from "src/hooks/useSettings";
 
 interface Column {
-  id: "id" | "visFunction" | "visDescription" | "visType" | "pageType" | "date";
+  id: "visFunction" | "visDescription" | "visType" | "pageType" | "date";
   label: string;
   minWidth?: number;
   align?: "right";
@@ -23,19 +24,11 @@ interface Column {
 }
 
 const columns: Column[] = [
-  { id: "id", label: "Link", minWidth: 170 },
-  { id: "visFunction", label: "VIS", minWidth: 100 },
-  { id: "visDescription", label: "VIS (Description)", minWidth: 100 },
-  { id: "visType", label: "VIS (Type)", minWidth: 100 },
-  { id: "pageType", label: "Page (Type)", minWidth: 100 },
+  { id: "visFunction", label: "VIS Function", minWidth: 100 },
+  { id: "visDescription", label: "Description", minWidth: 100 },
+  { id: "visType", label: "Type", minWidth: 100 },
+  { id: "pageType", label: "Page type", minWidth: 100 },
   { id: "date", label: "Date", minWidth: 100 },
-
-  //   label: "XX",
-  //   minWidth: 170,
-  //   align: "right",
-  //   format: (value: number) => value.toLocaleString("en-US"),  // OR
-  //   format: (value: number) => value.toFixed(2), // OR
-  //   format: (value: number) => value.toLocaleString("en-US"),
 ];
 
 interface Data {
@@ -45,6 +38,7 @@ interface Data {
 }
 
 interface PropagatedPageTableProps {
+  loading: boolean;
   data: Data[];
 }
 
@@ -63,31 +57,32 @@ const useStyles = makeStyles({
   },
 });
 
-// eslint-disable-next-line react/prop-types
-const PropagatedPageTable: FC<PropagatedPageTableProps> = ({ data = [] }) => {
+const PropagatedPageTable: FC<PropagatedPageTableProps> = (props) => {
   const { settings } = useSettings();
   const classes = useStyles();
   const router = useRouter();
-  const rows = data;
-
+  const rows = props.data;
   const [page, setPage] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(100);
 
   const handleChangePage = (_event: unknown, newPage: number) => {
     setPage(newPage);
+    Router.push(`/?page=${newPage}`);
+    console.log("PropagatedPageTable: newPage = ", newPage);
   };
 
   const handleChangeRowsPerPage = (
-    event: React.ChangeEvent<HTMLInputElement>
+    event: React.ChangeEvent<HTMLInputElement>,
   ) => {
     setRowsPerPage(+event.target.value);
     setPage(0);
+    console.log("PropagatedPageTable: setRowsPerPage = ", event.target.value);
   };
 
   return (
     <Paper className={classes.root}>
       <TableContainer className={classes.container}>
-        <Table stickyHeader aria-label="sticky table">
+        <Table stickyHeader size="small" aria-label="sticky table">
           <TableHead>
             <TableRow>
               {columns.map((column) => (
@@ -101,42 +96,50 @@ const PropagatedPageTable: FC<PropagatedPageTableProps> = ({ data = [] }) => {
               ))}
             </TableRow>
           </TableHead>
-          <TableBody>
-            {rows
-              .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-              .map((row) => (
-                <TableRow hover role="checkbox" tabIndex={-1} key={row.code}>
-                  {columns.map((column) => {
-                    const value = row[column.id];
-                    return (
-                      <TableCell key={column.id} align={column.align}>
-                        {/* {column.format &&
-                                      typeof value === "number"
-                                        ? column.format(value)
-                                        : value} */}
-
-                        {column.id === "id" ? (
-                          <IconButton
-                            color="primary"
-                            aria-label=""
-                            component="span"
-                            onClick={() => router.push(`/page/${row.id}`)}
-                          >
-                            <LinkIcon />
-                          </IconButton>
-                        ) : (
-                          value
-                        )}
-                      </TableCell>
-                    );
-                  })}
-                </TableRow>
-              ))}
-          </TableBody>
+          {props?.loading ? (
+            <Box
+              sx={{
+                display: "flex",
+                width: "100%",
+                height: 80,
+                alignItems: "center",
+                justifyContent: "center",
+              }}
+            >
+              <CircularProgress
+                variant="indeterminate"
+                color="primary"
+              ></CircularProgress>
+            </Box>
+          ) : (
+            <TableBody>
+              {props.data
+                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+                .map((row) => (
+                  <TableRow
+                    hover
+                    role="checkbox"
+                    tabIndex={-1}
+                    key={row.id}
+                    onClick={() =>
+                      router.push({ pathname: "/page", query: { id: row.id } })
+                    }
+                  >
+                    {columns.map((column) => {
+                      return (
+                        <TableCell key={column.id} align={column.align}>
+                          {row[column.id]}
+                        </TableCell>
+                      );
+                    })}
+                  </TableRow>
+                ))}
+            </TableBody>
+          )}
         </Table>
       </TableContainer>
       <TablePagination
-        rowsPerPageOptions={[100, 200, 300]}
+        rowsPerPageOptions={[5, 100, 200, 300]}
         component="div"
         count={rows.length}
         rowsPerPage={rowsPerPage}
