@@ -1,55 +1,48 @@
-import { useState } from 'react';
-import type { FC, ReactNode } from 'react';
-import { experimentalStyled } from '@material-ui/core/styles';
-import DashboardNavbar from './DashboardNavbar';
-import DashboardSidebar from './DashboardSidebar';
+import { useState } from "react";
+import type { FC, ReactNode } from "react";
+import { experimentalStyled } from "@mui/material/styles";
+import DashboardNavbar from "./DashboardNavbar";
+import DashboardSidebar from "./DashboardSidebar";
 
 interface DashboardLayoutProps {
   children?: ReactNode;
 }
 
-const DashboardLayoutRoot = experimentalStyled('div')(
-  ({ theme }) => (
-    {
-      backgroundColor: theme.palette.background.default,
-      display: 'flex',
-      height: '100%',
-      overflow: 'hidden',
-      width: '100%'
-    }
-  )
-);
+const DashboardLayoutRoot = experimentalStyled("div")(({ theme }) => ({
+  backgroundColor: theme.palette.background.default,
+  display: "flex",
+  height: "100%",
+  overflow: "hidden",
+  width: "100%",
+}));
 
-const DashboardLayoutWrapper = experimentalStyled('div')(
-  ({ theme }) => (
-    {
-      display: 'flex',
-      flex: '1 1 auto',
-      overflow: 'hidden',
-      paddingTop: '64px',
-      [theme.breakpoints.up('lg')]: {
-        paddingLeft: '280px'
-      }
-    }
-  )
-);
+const DashboardLayoutWrapper = experimentalStyled("div")(({ theme }) => ({
+  display: "flex",
+  flex: "1 1 auto",
+  overflow: "hidden",
+  paddingTop: "64px",
+  [theme.breakpoints.up("lg")]: {
+    paddingLeft: "280px",
+  },
+}));
 
-const DashboardLayoutContainer = experimentalStyled('div')({
-  display: 'flex',
-  flex: '1 1 auto',
-  overflow: 'hidden'
+const DashboardLayoutContainer = experimentalStyled("div")({
+  display: "flex",
+  flex: "1 1 auto",
+  overflow: "hidden",
 });
 
-const DashboardLayoutContent = experimentalStyled('div')({
-  flex: '1 1 auto',
-  height: '100%',
-  overflow: 'auto',
-  position: 'relative',
-  WebkitOverflowScrolling: 'touch'
+const DashboardLayoutContent = experimentalStyled("div")({
+  flex: "1 1 auto",
+  height: "100%",
+  overflow: "auto",
+  position: "relative",
+  WebkitOverflowScrolling: "touch",
 });
 
 const DashboardLayout: FC<DashboardLayoutProps> = ({ children }) => {
-  const [isSidebarMobileOpen, setIsSidebarMobileOpen] = useState<boolean>(false);
+  const [isSidebarMobileOpen, setIsSidebarMobileOpen] =
+    useState<boolean>(false);
 
   return (
     <DashboardLayoutRoot>
@@ -62,9 +55,7 @@ const DashboardLayout: FC<DashboardLayoutProps> = ({ children }) => {
       />
       <DashboardLayoutWrapper>
         <DashboardLayoutContainer>
-          <DashboardLayoutContent>
-            {children}
-          </DashboardLayoutContent>
+          <DashboardLayoutContent>{children}</DashboardLayoutContent>
         </DashboardLayoutContainer>
       </DashboardLayoutWrapper>
     </DashboardLayoutRoot>
