@@ -1,14 +1,10 @@
 import type { FC } from "react";
-import Link from "next/link";
 import PropTypes from "prop-types";
-import { AppBar, Box, IconButton, Toolbar } from "@material-ui/core";
-import { experimentalStyled } from "@material-ui/core/styles";
-import type { AppBarProps } from "@material-ui/core";
-import MenuIcon from "@material-ui/icons/Menu";
+import { AppBar, Box, IconButton, Toolbar } from "@mui/material";
+import { experimentalStyled } from "@mui/material/styles";
+import type { AppBarProps } from "@mui/material";
+import MenuIcon from "@mui/icons-material//Menu";
 import AccountPopover from "./AccountPopover";
-import Logo from "../Logo";
-import NotificationsPopover from "./NotificationsPopover";
-import ContentSearch from "./ContentSearch";
 
 interface DashboardNavbarProps extends AppBarProps {
   onSidebarMobileOpen?: () => void;
