@@ -1,34 +1,35 @@
-/* eslint-disable react/jsx-fragments */
-/* eslint-disable react/prop-types */
-/* eslint-disable arrow-body-style */
-
 import React, { FC } from "react";
-import { Theme, makeStyles } from "@material-ui/core/styles";
-import List from "@material-ui/core/List";
-import ListItem from "@material-ui/core/ListItem";
-import Divider from "@material-ui/core/Divider";
-import ListItemText from "@material-ui/core/ListItemText";
-import ListItemAvatar from "@material-ui/core/ListItemAvatar";
-import Avatar from "@material-ui/core/Avatar";
-import Typography from "@material-ui/core/Typography";
-import { Chip, Box } from "@material-ui/core";
-import ImageIcon from "@material-ui/icons/Image";
+import { makeStyles } from "@mui/styles";
+import {
+  Box,
+  Typography,
+  Avatar,
+  Divider,
+  List,
+  ListItem,
+  ListItemText,
+  ListItemAvatar,
+} from "@mui/material";
+import ImageIcon from "@mui/icons-material/Image";
+import Link from "next/link";
 
-const useStyles = makeStyles((theme: Theme) => ({
+const API_JS = process.env.NEXT_PUBLIC_API_JS;
+
+const useStyles = makeStyles(() => ({
   root: {
     width: "100%",
     maxWidth: "100%",
-    backgroundColor: theme.palette.background.paper,
+    // backgroundColor: theme.palette.background.paper,
   },
   inline: {
     display: "inline",
   },
   chip: {
-    margin: theme.spacing(0.5),
+    // margin: theme.spacing(0.5),
   },
   large: {
-    width: theme.spacing(5),
-    height: theme.spacing(5),
+    // width: theme.spacing(5),
+    // height: theme.spacing(5),
     margin: "5px 10px",
   },
 }));
@@ -39,21 +40,11 @@ interface SearchResultProps {
 
 const SearchResultView: FC<SearchResultProps> = ({ data = [] }) => {
   const classes = useStyles();
- 
-  // TODO: Use real images from backend
-  const list = [
-    "605e64ccdfb1d977d34aa3cc.png",
-    "609728d27d47ae21406735bd.png",
-    "60ecc0f3beb7791f01bebe49.png",
-    "61006ed44fef9b1f276003de.png",
-    "61031507be36153857a3de37.png",
-    "608dd7dbd651fc539ce11801.png",
-    "60ad693df52d2d641f4e45b9.png",
-    "61006c9842248f1ef21219b1.png",
-    "610314efc50719383382a6a2.png",
-  ];
 
-  console.log("SearchResultView: data =", data);
+  // FIXME: Remove (useEffect makes sure that data is not logged during SSR)
+  React.useEffect(() => {
+    console.log("SearchResultView: data =", data);
+  }, [data]);
 
   return (
     <Box
@@ -64,17 +55,27 @@ const SearchResultView: FC<SearchResultProps> = ({ data = [] }) => {
       m={1}
     >
       <List className={classes.root}>
-        {data.map((d: any) => (
-          <>
-            <ListItem 
-              alignItems="flex-start"
-            >
-              <ListItemAvatar>
-                <Avatar variant="square" className={classes.large} src={`/static/mock-images/${list[Math.floor(Math.random() * list.length)]}`}>
-                </Avatar>
-              </ListItemAvatar>
+        {data.map((dataRecord, index) => (
+          <React.Fragment key={index}>
+            <ListItem alignItems="flex-start">
+              <a
+                href={`/page?id=${dataRecord.id}`}
+                target="_blank"
+                rel="noreferrer"
+              >
+                <ListItemAvatar>
+                  <Avatar
+                    variant="square"
+                    className={classes.large}
+                    src={`data:image/jpeg;base64,${dataRecord.thumbnail}`}
+                  >
+                    <ImageIcon />
+                  </Avatar>
+                </ListItemAvatar>
+              </a>
+
               <ListItemText
-                primary={data?.title}
+                primary={dataRecord?.title}
                 secondary={
                   <React.Fragment>
                     <Typography
@@ -83,10 +84,10 @@ const SearchResultView: FC<SearchResultProps> = ({ data = [] }) => {
                       className={classes.inline}
                       color="textPrimary"
                     >
-                      {d.title}
+                      {/* {dataRecord.title} */}
                       <Divider orientation="vertical" />
                     </Typography>
-                    {d.visDescription}
+                    {dataRecord.visDescription}
                     <Divider orientation="vertical" />
 
                     {/* {d.keywords.map((k: string) => (
@@ -98,7 +99,7 @@ const SearchResultView: FC<SearchResultProps> = ({ data = [] }) => {
             </ListItem>
 
             <Divider variant="inset" component="li" />
-          </>
+          </React.Fragment>
         ))}
       </List>
     </Box>
