@@ -1,23 +1,11 @@
-import type { ReactElement } from "react";
-import Link from "next/link";
+import { ReactElement } from "react";
 import { Helmet } from "react-helmet-async";
-import {
-  Box,
-  Card,
-  CardContent,
-  Container,
-  Divider,
-  Typography,
-  Button,
-} from "@material-ui/core";
+import { Box, Card, CardContent, Container, Divider } from "@mui/material";
 import { LoginJWT } from "src/components/auth/login";
-import Logo from "src/components/Logo";
-import useAuth from "src/hooks/useAuth";
 import MainLayout from "src/components/main-layout/MainLayout";
+import LoginGitHub from "src/components/auth/login/LoginGitHub";
 
 const Login = () => {
-  const { platform } = useAuth() as any;
-
   return (
     <>
       <Helmet>
@@ -39,14 +27,7 @@ const Login = () => {
               justifyContent: "center",
               mb: 8,
             }}
-          >
-            <Logo
-              sx={{
-                height: 60,
-                width: 60,
-              }}
-            />
-          </Box>
+          ></Box>
           <Card>
             <CardContent
               sx={{
@@ -57,49 +38,24 @@ const Login = () => {
             >
               <Box
                 sx={{
-                  alignItems: "center",
-                  display: "flex",
-                  justifyContent: "space-between",
-                  mb: 3,
+                  flexGrow: 1,
+                  mt: 3,
                 }}
               >
-                <div>
-                  <Typography color="textPrimary" gutterBottom variant="h4">
-                    Log in
-                  </Typography>
-                  <Typography color="textSecondary" variant="body2">
-                    Log in on the RAMPVIS platform
-                  </Typography>
-                </div>
+                <LoginGitHub />
               </Box>
+
+              <Divider sx={{ my: 3 }} />
+
               <Box
                 sx={{
                   flexGrow: 1,
                   mt: 3,
                 }}
               >
-                {platform === "JWT" && <LoginJWT />}
+                <LoginJWT />
               </Box>
               <Divider sx={{ my: 3 }} />
-
-              <Box
-                sx={{
-                  display: "flex",
-                  justifyContent: "center",
-                  mt: 6,
-                }}
-              >
-                {/* 
-                <Button
-                  color="primary"
-                  component={Link}
-                  href="/"
-                  variant="outlined"
-                >
-                  Go to Home
-                </Button> 
-              */}
-              </Box>
             </CardContent>
           </Card>
         </Container>
