@@ -1,14 +1,7 @@
 /** @type {import('next').NextConfig} */
 module.exports = {
   reactStrictMode: true,
-  async redirects() {
-      return [
-          {
-              source: '/',
-              destination: '/search',
-              permanent: false,
-          }
-      ]
-  }
-
-}
+  typescript: {
+    ignoreBuildErrors: true,
+  },
+};
