@@ -7,7 +7,7 @@ import PropTypes from "prop-types";
 import { orange, grey } from "@material-ui/core/colors";
 import ShowHideGuard from "src/components/auth/guards/ShowHideGuard";
 import { apiService } from "src/utils/apiService";
-import useAuth from "src/hooks/useAuth";
+// import useAuth from "src/hooks/useAuth";
 
 const useStyles = makeStyles((theme) => ({
   bookmarkedStyle: {
@@ -23,21 +23,24 @@ const useStyles = makeStyles((theme) => ({
 const Bookmark: FC<any> = ({ pageId }) => {
   const classes = useStyles();
   const [isBookmarked, setBookmark] = useState<boolean>(false);
-  const { user } = useAuth();
+  // const { user } = useAuth();
 
-  if (user?.bookmarks?.includes(pageId)) {
-    setBookmark(true); 
-  }
-  console.log("Bookmark: isBookmarked = ", isBookmarked);
+  // if (user?.bookmarks?.includes(pageId)) {
+  //   setBookmark(true);
+  // }
+  // console.log("Bookmark: isBookmarked = ", isBookmarked);
 
   const onClickBookmark = async () => {
     console.log("Bookmark: !isBookmarked = ", !isBookmarked);
-    const res = await apiService.post<any>(`/me/bookmark`, {pageId, status: !isBookmarked});
+    const res = await apiService.post<any>(`/me/bookmark`, {
+      pageId,
+      status: !isBookmarked,
+    });
     // if (user?.bookmarks?.includes(pageId)) {
-    //   setBookmark(true); 
+    //   setBookmark(true);
     // }
     // TODO update user
-    console.log(res)
+    console.log("Bookmark: res = ", res);
   };
 
   return (
@@ -45,7 +48,9 @@ const Bookmark: FC<any> = ({ pageId }) => {
       <IconButton
         aria-label="bookmark"
         onClick={() => onClickBookmark()}
-        className={ isBookmarked ? classes.bookmarkedStyle : classes.unBookmarkedStyle }
+        className={
+          isBookmarked ? classes.bookmarkedStyle : classes.unBookmarkedStyle
+        }
       >
         <BookmarkIcon fontSize="inherit" />
       </IconButton>
