@@ -1,5 +1,5 @@
 import type { FC, ReactNode } from "react";
-import { Navigate } from "react-router-dom";
+import { useRouter } from "next/router";
 import PropTypes from "prop-types";
 import useAuth from "src/hooks/useAuth";
 
@@ -9,9 +9,11 @@ interface GuestGuardProps {
 
 const GuestGuard: FC<GuestGuardProps> = ({ children }) => {
   const { isAuthenticated } = useAuth();
+  const router = useRouter();
 
   if (isAuthenticated) {
-    return <Navigate to="/pages/example" />;
+    router.push("/pages/example");
+    return <></>;
   }
 
   return <>{children}</>;
