@@ -151,8 +151,7 @@ const ContentSearch: FC = () => {
                       <>
                         {results.map((result, i) => (
                           <Box key={i} sx={{ mb: 2 }}>
-                            {
-                            /* <Link
+                            {/* <Link
                               color="textPrimary"
                               component={RouterLink}
                               to="/dashboard"
