@@ -1,12 +1,12 @@
-import { useEffect, FC } from 'react';
+import { useEffect, FC } from "react";
 // import { useLocation } from 'react-router-dom';
 import Link from "next/link";
-import { useRouter } from 'next/router'
-import PropTypes from 'prop-types';
-import { Box, Chip, Drawer } from '@material-ui/core';
-import type { Theme } from '@material-ui/core';
-import useMediaQuery from '@material-ui/core/useMediaQuery';
-import Logo from 'src/components/Logo';
+import { useRouter } from "next/router";
+import PropTypes from "prop-types";
+import { Box, Chip, Drawer } from "@material-ui/core";
+import type { Theme } from "@material-ui/core";
+import useMediaQuery from "@material-ui/core/useMediaQuery";
+import Logo from "src/components/Logo";
 
 interface MainSidebarProps {
   onMobileClose: () => void;
@@ -16,8 +16,8 @@ interface MainSidebarProps {
 const MainSidebar: FC<MainSidebarProps> = (props) => {
   const { onMobileClose, openMobile } = props;
   // const location = useLocation();
-  const router = useRouter()
-  const lgUp = useMediaQuery((theme: Theme) => theme?.breakpoints.up('lg'));
+  const router = useRouter();
+  const lgUp = useMediaQuery((theme: Theme) => theme?.breakpoints.up("lg"));
 
   useEffect(() => {
     if (openMobile && onMobileClose) {
@@ -33,18 +33,18 @@ const MainSidebar: FC<MainSidebarProps> = (props) => {
       variant="temporary"
       PaperProps={{
         sx: {
-          backgroundColor: 'background.default',
-          width: 256
-        }
+          backgroundColor: "background.default",
+          width: 256,
+        },
       }}
     >
       <Box
         sx={{
-          alignItems: 'flex-start',
-          display: 'flex',
-          flexDirection: 'column',
-          height: '100%',
-          p: 2
+          alignItems: "flex-start",
+          display: "flex",
+          flexDirection: "column",
+          height: "100%",
+          p: 2,
         }}
       >
         <Link href="/">
@@ -52,9 +52,9 @@ const MainSidebar: FC<MainSidebarProps> = (props) => {
         </Link>
         <Box
           sx={{
-            display: 'flex',
+            display: "flex",
             pb: 2,
-            pt: 3
+            pt: 3,
           }}
         >
           <Chip
@@ -64,11 +64,10 @@ const MainSidebar: FC<MainSidebarProps> = (props) => {
             sx={{
               maxHeight: 20,
               ml: 1,
-              mr: 2
+              mr: 2,
             }}
           />
         </Box>
-       
       </Box>
     </Drawer>
   );
@@ -76,7 +75,7 @@ const MainSidebar: FC<MainSidebarProps> = (props) => {
 
 MainSidebar.propTypes = {
   onMobileClose: PropTypes.func,
-  openMobile: PropTypes.bool
+  openMobile: PropTypes.bool,
 };
 
 export default MainSidebar;
