@@ -1,11 +1,11 @@
-import { createContext, useEffect, useState } from 'react';
-import type { FC, ReactNode } from 'react';
-import PropTypes from 'prop-types';
-import { THEMES } from '../constants';
+import { createContext, useEffect, useState } from "react";
+import type { FC, ReactNode } from "react";
+import PropTypes from "prop-types";
+import { THEMES } from "../constants";
 
 interface Settings {
   compact?: boolean;
-  direction?: 'ltr' | 'rtl';
+  direction?: "ltr" | "rtl";
   responsiveFontSizes?: boolean;
   roundedCorners?: boolean;
   theme?: string;
@@ -22,29 +22,29 @@ interface SettingsProviderProps {
 
 const initialSettings: Settings = {
   compact: true,
-  direction: 'ltr',
+  direction: "ltr",
   responsiveFontSizes: true,
   roundedCorners: true,
-  theme: THEMES.LIGHT
+  theme: THEMES.LIGHT,
 };
 
 export const restoreSettings = (): Settings | null => {
   let settings = null;
 
   try {
-    const storedData: string | null = window.localStorage.getItem('settings');
+    const storedData: string | null = window.localStorage.getItem("settings");
 
     if (storedData) {
       settings = JSON.parse(storedData);
     } else {
       settings = {
         compact: true,
-        direction: 'ltr',
+        direction: "ltr",
         responsiveFontSizes: true,
         roundedCorners: true,
-        theme: window.matchMedia('(prefers-color-scheme: dark)').matches
+        theme: window.matchMedia("(prefers-color-scheme: dark)").matches
           ? THEMES.DARK
-          : THEMES.LIGHT
+          : THEMES.LIGHT,
       };
     }
   } catch (err) {
@@ -57,12 +57,12 @@ export const restoreSettings = (): Settings | null => {
 };
 
 export const storeSettings = (settings: Settings): void => {
-  window.localStorage.setItem('settings', JSON.stringify(settings));
+  window.localStorage.setItem("settings", JSON.stringify(settings));
 };
 
 const SettingsContext = createContext<SettingsContextValue>({
   settings: initialSettings,
-  saveSettings: () => {}
+  saveSettings: () => {},
 });
 
 export const SettingsProvider: FC<SettingsProviderProps> = (props) => {
@@ -86,7 +86,7 @@ export const SettingsProvider: FC<SettingsProviderProps> = (props) => {
     <SettingsContext.Provider
       value={{
         settings,
-        saveSettings
+        saveSettings,
       }}
     >
       {children}
@@ -95,7 +95,7 @@ export const SettingsProvider: FC<SettingsProviderProps> = (props) => {
 };
 
 SettingsProvider.propTypes = {
-  children: PropTypes.node.isRequired
+  children: PropTypes.node.isRequired,
 };
 
 export const SettingsConsumer = SettingsContext.Consumer;
