@@ -1,14 +1,14 @@
-import { useEffect } from 'react';
+import { useEffect } from "react";
 // import { useLocation } from 'react-router-dom';
-import { useRouter } from 'next/router'
+import { useRouter } from "next/router";
 
 const useScrollReset = (): null => {
   // const location = useLocation();
-  const router = useRouter()
+  const router = useRouter();
 
   useEffect(() => {
     window.scrollTo(0, 0);
-  }, [router.asPath] );
+  }, [router.asPath]);
 
   return null;
 };
