@@ -21,139 +21,148 @@ import * as d3 from "d3";
  * Stacked Area Chart.
  */
 export function stackedAreaChart() {
+  /**
+   * Visual configs.
+   */
+  let margin = { top: 10, right: 10, bottom: 10, left: 10 };
+
+  let visWidth = 960,
+    visHeight = 600, // Size of the visualization, including margins
+    width,
+    height; // Size of the main content, excluding margins
+
+  /**
+   * Accessors.
+   */
+  let time = (d) => d.time;
+
+  /**
+   * Data binding to DOM elements.
+   */
+  let data;
+
+  /**
+   * DOM.
+   */
+  let visContainer, itemContainer, xAxisContainer, yAxisContainer;
+
+  /**
+   * D3.
+   */
+  const xScale = d3.scaleTime(),
+    yScale = d3.scaleLinear(),
+    xAxis = d3.axisBottom().scale(xScale).ticks(d3.timeWeek),
+    yAxis = d3.axisLeft().scale(yScale).ticks(5),
+    area = d3
+      .area()
+      .x((d) => xScale(time(d.data)))
+      .y0((d) => yScale(d[0]))
+      .y1((d) => yScale(d[1]));
+  let colorScale;
+
+  /**
+   * Main entry of the module.
+   */
+  function module(selection) {
+    selection.each(function (_data) {
+      if (!this.visInitialized) {
+        visContainer = d3
+          .select(this)
+          .append("g")
+          .attr("class", "pv-stacked-area-chart");
+        xAxisContainer = visContainer.append("g").attr("class", "x-axis");
+        yAxisContainer = visContainer.append("g").attr("class", "y-axis");
+        itemContainer = visContainer.append("g").attr("class", "items");
+
+        this.visInitialized = true;
+      }
+
+      data = _data;
+      update();
+    });
+  }
+
+  /**
+   * Updates the visualization when data or display attributes changes.
+   */
+  function update() {
     /**
-     * Visual configs.
+     * Display area.
      */
-    let margin = { top: 10, right: 10, bottom: 10, left: 10 };
-
-    let visWidth = 960, visHeight = 600, // Size of the visualization, including margins
-        width, height; // Size of the main content, excluding margins
-
-    /**
-     * Accessors.
-     */
-    let time = d => d.time;
-
-    /**
-     * Data binding to DOM elements.
-     */
-    let data;
-
-    /**
-     * DOM.
-     */
-    let visContainer,
-        itemContainer,
-        xAxisContainer,
-        yAxisContainer;
-
-    /**
-     * D3.
-     */
-    const xScale = d3.scaleTime(),
-        yScale = d3.scaleLinear(),
-        xAxis = d3.axisBottom().scale(xScale).ticks(d3.timeWeek),
-        yAxis = d3.axisLeft().scale(yScale).ticks(5),
-        area = d3.area()
-            .x(d => xScale(time(d.data)))
-            .y0(d => yScale(d[0]))
-            .y1(d => yScale(d[1]));
-    let colorScale;
-
-    /**
-     * Main entry of the module.
-     */
-    function module(selection) {
-        selection.each(function(_data) {
-            if (!this.visInitialized) {
-                visContainer = d3.select(this).append('g').attr('class', 'pv-stacked-area-chart');
-                xAxisContainer = visContainer.append('g').attr('class', 'x-axis');
-                yAxisContainer = visContainer.append('g').attr('class', 'y-axis');
-                itemContainer = visContainer.append('g').attr('class', 'items');
-
-                this.visInitialized = true;
-            }
-
-            data = _data;
-            update();
-        });
-    }
-
-    /**
-     * Updates the visualization when data or display attributes changes.
-     */
-    function update() {
-        /**
-         * Display area.
-         */
-        width = visWidth - margin.left - margin.right;
-        height = visHeight - margin.top - margin.bottom;
-        visContainer.attr('transform', 'translate(' + margin.left + ',' + margin.top + ')');
-        xAxisContainer.attr('transform', 'translate(0,' + height + ')');
-
-        /**
-         * Computation.
-         */
-        const series = d3.stack().keys(data.columns)(data);
-        
-        xScale.domain(d3.extent(data, time))
-            .range([0, width]);
-        yScale.domain([0, d3.max(series, d => d3.max(d, d => d[1]))]).nice()
-            .range([height, 0]);
-        
-        /**
-         * Draw.
-         */
-        itemContainer.selectAll('path')
-            .data(series)
-            .join('path')
-                .attr('fill', ({ key }) => colorScale(key))
-                .attr('d', area)
-            .append('title')
-                .text(({ key }) => key);
-
-        xAxisContainer.call(xAxis)
-            .selectAll('text')
-            .attr('transform', 'rotate(-45)')
-            .style('text-anchor', 'end')
-        yAxisContainer.call(yAxis);
-    }
+    width = visWidth - margin.left - margin.right;
+    height = visHeight - margin.top - margin.bottom;
+    visContainer.attr(
+      "transform",
+      "translate(" + margin.left + "," + margin.top + ")",
+    );
+    xAxisContainer.attr("transform", "translate(0," + height + ")");
 
     /**
-     * Sets/gets the margin of the visualization.
+     * Computation.
      */
-    module.margin = function(value) {
-        if (!arguments.length) return margin;
-        margin = value;
-        return this;
-    };
+    const series = d3.stack().keys(data.columns)(data);
+
+    xScale.domain(d3.extent(data, time)).range([0, width]);
+    yScale
+      .domain([0, d3.max(series, (d) => d3.max(d, (d) => d[1]))])
+      .nice()
+      .range([height, 0]);
 
     /**
-     * Sets/gets the width of the visualization.
+     * Draw.
      */
-    module.width = function(value) {
-        if (!arguments.length) return visWidth;
-        visWidth = value;
-        return this;
-    };
+    itemContainer
+      .selectAll("path")
+      .data(series)
+      .join("path")
+      .attr("fill", ({ key }) => colorScale(key))
+      .attr("d", area)
+      .append("title")
+      .text(({ key }) => key);
 
-    /**
-     * Sets/gets the height of the visualization.
-     */
-    module.height = function(value) {
-        if (!arguments.length) return visHeight;
-        visHeight = value;
-        return this;
-    };
+    xAxisContainer
+      .call(xAxis)
+      .selectAll("text")
+      .attr("transform", "rotate(-45)")
+      .style("text-anchor", "end");
+    yAxisContainer.call(yAxis);
+  }
 
-    /**
-     * Sets/gets the color scale.
-     */
-    module.colorScale = function(value) {
-        if (!arguments.length) return colorScale;
-        colorScale = value;
-        return this;
-    };
+  /**
+   * Sets/gets the margin of the visualization.
+   */
+  module.margin = function (value) {
+    if (!arguments.length) return margin;
+    margin = value;
+    return this;
+  };
 
-    return module;
-};
+  /**
+   * Sets/gets the width of the visualization.
+   */
+  module.width = function (value) {
+    if (!arguments.length) return visWidth;
+    visWidth = value;
+    return this;
+  };
+
+  /**
+   * Sets/gets the height of the visualization.
+   */
+  module.height = function (value) {
+    if (!arguments.length) return visHeight;
+    visHeight = value;
+    return this;
+  };
+
+  /**
+   * Sets/gets the color scale.
+   */
+  module.colorScale = function (value) {
+    if (!arguments.length) return colorScale;
+    colorScale = value;
+    return this;
+  };
+
+  return module;
+}
