@@ -30,8 +30,6 @@ import "src/lib/vis/css/default-dashboard.css";
 import "src/lib/vis/css/overview-top-level-screen-a.css";
 import "src/lib/vis/css/portal.css";
 import "src/lib/vis/css/pv-legend.css";
-import "src/lib/vis/css/vis-example.css";
-
 
 type NextPageWithLayout = NextPage & {
   getLayout?: (page: ReactElement) => ReactNode;
@@ -77,9 +75,9 @@ function MyApp({ Component, pageProps }: AppPropsWithLayout) {
               <SettingsProvider>
                 <AuthProviderJWT>
                   <ThemeProvider theme={theme}>
-                      <CssBaseline />
-                      <Toaster position="top-center" />
-                      {getLayout(<Component {...pageProps} />)}
+                    <CssBaseline />
+                    <Toaster position="top-center" />
+                    {getLayout(<Component {...pageProps} />)}
                   </ThemeProvider>
                 </AuthProviderJWT>
               </SettingsProvider>
@@ -90,10 +88,5 @@ function MyApp({ Component, pageProps }: AppPropsWithLayout) {
     </>
   );
 }
-
-MyApp.getInitialProps = async (appContext: AppContext) => {
-  const appProps = await App.getInitialProps(appContext);
-  return { ...appProps };
-};
 
 export default MyApp;
