@@ -1,29 +1,14 @@
-/* eslint-disable arrow-body-style */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-
-import { ReactElement, useCallback, useEffect, useState } from "react";
+import { ReactElement } from "react";
 import { Helmet } from "react-helmet-async";
 import { useRouter } from "next/router";
 import _ from "lodash";
-import {
-  Avatar,
-  Box,
-  Card,
-  CardContent,
-  CardHeader,
-  Container,
-  Grid,
-  IconButton,
-  Typography,
-} from "@material-ui/core";
 import DashboardLayout from "src/components/dashboard-layout/DashboardLayout";
-import useSettings from "src/hooks/useSettings";
 import UnderConstruction from "src/components/errors/UnderConstruction";
+import { GetStaticProps, GetStaticPaths, NextPage } from "next";
 
 const MyPortal = () => {
-  const { settings } = useSettings();
   const router = useRouter();
-  const { country } = router.query;
+  const country = `${router.query.country}`;
 
   return (
     <>
@@ -38,6 +23,22 @@ const MyPortal = () => {
 
 MyPortal.getLayout = function getLayout(page: ReactElement) {
   return <DashboardLayout>{page}</DashboardLayout>;
+};
+
+export const getStaticProps: GetStaticProps = () => {
+  return { props: {} };
+};
+
+export const getStaticPaths: GetStaticPaths = () => {
+  return {
+    paths: [
+      "/country/england",
+      "/country/northern-ireland",
+      "/country/scotland",
+      "/country/wales",
+    ],
+    fallback: false,
+  };
 };
 
 export default MyPortal;
