@@ -4,7 +4,6 @@
 import axios from "axios";
 
 const API_JS = process.env.NEXT_PUBLIC_API_JS;
-const API_TS = process.env.NEXT_PUBLIC_API_TS;
 
 export class ApiService {
   private endpoint: string = API_JS;
@@ -18,31 +17,43 @@ export class ApiService {
       url = `/${url}`;
     }
 
-    console.log(API_JS, this.endpoint, url)
+    console.log(API_JS, this.endpoint, url);
     return this.endpoint + url;
   }
 
   getAxiosInstance() {
     return axios.create({
       // timeout: 60 * 1 * 1000, // 1 minute
-      headers: { Authorization: `Bearer ${window.localStorage.getItem("accessToken")}` },
+      headers: {
+        Authorization: `Bearer ${window.localStorage.getItem("accessToken")}`,
+      },
     });
   }
 
   // HTTP Methods
 
   async get<T extends any>(url: string): Promise<T> {
-    console.log("ApiService:get: url = ", url, "endpoint = ", this.getEndpoint(url))
+    console.log(
+      "ApiService:get: url = ",
+      url,
+      "endpoint = ",
+      this.getEndpoint(url),
+    );
     const res = await this.getAxiosInstance().get(this.getEndpoint(url));
     return res.data;
   }
 
   async post<T extends any>(url: string, body: T): Promise<T> {
-    console.log("ApiService:post: url = ", url, "endpoint = ", this.getEndpoint(url))
+    console.log(
+      "ApiService:post: url = ",
+      url,
+      "endpoint = ",
+      this.getEndpoint(url),
+    );
     const res = await this.getAxiosInstance().post(this.getEndpoint(url), body);
     return res.data;
   }
-  
+
   async delete<T extends any>(url: string): Promise<T> {
     const res = await this.getAxiosInstance().delete(this.getEndpoint(url));
     return res.data;
