@@ -1,3 +1,4 @@
-const wait = (time: number): Promise<void> => new Promise((res) => setTimeout(res, time));
+const wait = (time: number): Promise<void> =>
+  new Promise((res) => setTimeout(res, time));
 
 export default wait;
